Initial implementation of model import from Magica_voxel format
