+++ Cargo.toml
@@ -9,11 +9,14 @@ license = "MIT OR Apache-2.0"
 default = []
 raytracing = ["dep:image", "dep:show-image"]
 serialization = ["dep:serde"]
+magica_voxel = ["dep:dot_vox"]
 bevy_wgpu = ["raytracing", "dep:bevy"]
 
 [dependencies]
 serde = { version = "1.0.183", features = ["derive"], optional = true }
 bendy = { git = "https://github.com/davids91/bendy.git" , features = ["std", "serde"]}
+dot_vox = { version = "5.1.1", optional = true }
+
 # for example cpu_render
 image = { version = "0.25.1", optional = true }
 show-image = { version = "0.14.0", optional = true }
