+++ assets/shaders/viewport_render.wgsl
@@ -468,10 +468,7 @@ fn get_by_ray(ray_: Line) -> OctreeRayIntersection{
         );
         node_stack_i = 1;
     }
-
-    var i = 0.;
     while(0 < node_stack_i && node_stack_i < max_depth) {
-        i += 1.;
         var current_bounds = node_stack[node_stack_i - 1].bounds;
         var current_node = nodes[node_stack[node_stack_i - 1].node]; //!NOTE: should be const, but then it can not be indexed dynamically
         var target_octant = node_stack[node_stack_i - 1].target_octant;
