+++ src/octree/convert/mod.rs
@@ -1,0 +1,4 @@
+mod bytecode;
+
+#[cfg(feature = "magica_voxel")]
+mod magicavoxel;
