+++ src/octree/types.rs
@@ -55,48 +55,15 @@ pub trait VoxelData {
     fn clear(&mut self);
 }
 
-impl VoxelData for Albedo {
-    fn new(color: Albedo, _user_data: u32) -> Self {
-        color
-    }
-
-    fn albedo(&self) -> Albedo {
-        *self
-    }
-
-    fn user_data(&self) -> u32 {
-        0u32
-    }
-
-    fn clear(&mut self) {
-        self.r = 0;
-        self.r = 0;
-        self.b = 0;
-        self.a = 0;
-    }
-}
-
-impl From<u32> for Albedo {
-    fn from(value: u32) -> Self {
-        let a = (value & 0x000000FF) as u8;
-        let b = ((value & 0x0000FF00) >> 8) as u8;
-        let g = ((value & 0x00FF0000) >> 16) as u8;
-        let r = ((value & 0xFF000000) >> 24) as u8;
-
-        Albedo::default()
-            .with_red(r)
-            .with_green(g)
-            .with_blue(b)
-            .with_alpha(a)
-    }
-}
-
 /// Sparse Octree of Nodes, where each node contains a brick of voxels.
 /// A Brick is a 3 dimensional matrix, each element of it containing a voxel.
 /// A Brick can be indexed directly, as opposed to the octree which is essentially a
 /// tree-graph where each node has 8 children.
 #[cfg_attr(feature = "serialization", derive(Serialize))]
-pub struct Octree<T: Default + Clone + VoxelData, const DIM: usize = 1> {
+pub struct Octree<T, const DIM: usize = 1>
+where
+    T: Default + Clone + VoxelData,
+{
     pub auto_simplify: bool,
     pub(in crate::octree) octree_size: u32,
     pub(in crate::octree) nodes: ObjectPool<NodeContent<T, DIM>>,
