+++ src/octree/update.rs
@@ -13,7 +13,7 @@ use crate::spatial::{
 
 impl<T, const DIM: usize> Octree<T, DIM>
 where
-    T: Default + Eq + Clone + Copy + VoxelData,
+    T: Default + PartialEq + Clone + Copy + VoxelData,
 {
     /// Inserts the given data into the octree into the intended voxel position
     pub fn insert(&mut self, position: &V3c<u32>, data: T) -> Result<(), OctreeError> {
